requirements.txt
# Core dependencies
git
cmake
pillow
numpy
opencv-python
torch
torchvision
torchaudio
flask
pydantic

# Machine Learning & Image Processing
insightface>=0.7.3
onnxruntime
git+https://github.com/davisking/dlib.git
scikit-image
scipy
lpips

# Utilities
protobuf
addict
future
lmdb
pyyaml
requests
tqdm
yapf
gdown

# Web & API
streamlit
fastapi
uvicorn
python-multipart

# Development & Logging
tb-nightly